Rename multilanguage objects.
