--- framework/src/Volo.Abp.MultiLingualObject/Volo/Abp/MultiLingualObject/IHasMultiLingual.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Volo.Abp.MultiLingualObject
-{
-    public interface IHasMultiLingual<TTranslation>
-        where TTranslation : class, IMultiLingualTranslation
-    {
-        ICollection<TTranslation> Translations { get; set; }
-    }
-}
