--- framework/src/Volo.Abp.MultiLingualObject/Volo/Abp/MultiLingualObject/IMultiLingualTranslation.cs
@@ -1,7 +1,0 @@
-﻿namespace Volo.Abp.MultiLingualObject
-{
-    public interface IMultiLingualTranslation
-    {
-        string Language { get; set; }
-    }
-}
