+++ framework/src/Volo.Abp.MultiLingualObjects/Volo/Abp/MultiLingualObjects/IMultiLingualObject.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Volo.Abp.MultiLingualObjects
+{
+    public interface IMultiLingualObject<TTranslation>
+        where TTranslation : class, IObjectTranslation
+    {
+        ICollection<TTranslation> Translations { get; set; }
+    }
+}
