+++ framework/src/Volo.Abp.MultiLingualObjects/Volo/Abp/MultiLingualObjects/IObjectTranslation.cs
@@ -1,0 +1,7 @@
+﻿namespace Volo.Abp.MultiLingualObjects
+{
+    public interface IObjectTranslation
+    {
+        string Language { get; set; }
+    }
+}
