+++ framework/test/Volo.Abp.AutoMapper.Tests/Volo/Abp/AutoMapper/AutoMapperTestModule.cs
@@ -1,5 +1,4 @@
 ﻿using Volo.Abp.Modularity;
-using Volo.Abp.MultiLingualObject;
 using Volo.Abp.ObjectExtending;
 
 namespace Volo.Abp.AutoMapper
