--- framework/test/Volo.Abp.MultiLingualObject.Tests/Volo/Abp/MultiLingualObject/TestObjects/MultiLingualBookTranslation.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Volo.Abp.MultiLingualObject.TestObjects
-{
-    public class MultiLingualBookTranslation : IMultiLingualTranslation
-    {
-        public string Name { get; set; }
-
-        public string Language { get; set; }
-    }
-}
