+++ framework/test/Volo.Abp.MultiLingualObjects.Tests/Volo/Abp/MultiLingualObjects/AbpMultiLingualObjectsTestModule.cs
@@ -1,21 +1,20 @@
-﻿using Autofac.Extensions.DependencyInjection;
-using Volo.Abp.Autofac;
+﻿using Volo.Abp.Autofac;
 using Volo.Abp.Localization;
 using Volo.Abp.Modularity;
 using Volo.Abp.ObjectMapping;
 using Volo.Abp.Settings;
 
-namespace Volo.Abp.MultiLingualObject
+namespace Volo.Abp.MultiLingualObjects
 {
     [DependsOn(
         typeof(AbpAutofacModule),
         typeof(AbpLocalizationModule),
         typeof(AbpSettingsModule),
         typeof(AbpObjectMappingModule),
-        typeof(AbpMultiLingualObjectModule),
+        typeof(AbpMultiLingualObjectsModule),
         typeof(AbpTestBaseModule)
     )]
-    public class AbpMultiLingualObjectTestModule : AbpModule
+    public class AbpMultiLingualObjectsTestModule : AbpModule
     {
     }
 }
