+++ framework/test/Volo.Abp.MultiLingualObjects.Tests/Volo/Abp/MultiLingualObjects/TestObjects/MultiLingualBookTranslation.cs
@@ -1,0 +1,9 @@
+﻿namespace Volo.Abp.MultiLingualObjects.TestObjects
+{
+    public class MultiLingualBookTranslation : IObjectTranslation
+    {
+        public string Name { get; set; }
+
+        public string Language { get; set; }
+    }
+}
